+++ lib/pages/Home/Widgets/dashboard_menu.dart
@@ -1,11 +1,7 @@
-import 'dart:math';
-
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/screenutil.dart';
-import 'package:gdziemojhajsapp/logic/Constants/ReceiptSortTypeEnum.dart';
 import 'package:gdziemojhajsapp/pages/Home/Widgets/receipt_widgets.dart';
-import 'package:gdziemojhajsapp/pages/Home/home_screen.dart';
 import 'package:gdziemojhajsapp/pages/Receipt/createReceipt.dart';
 import 'package:percent_indicator/linear_percent_indicator.dart';
 import 'package:unicorndial/unicorndial.dart';
@@ -207,108 +203,4 @@ appBarWidget(notifyParent) {
       ),
     ),
   );
-}
-
-class SortReceiptsBar extends StatefulWidget {
-  SortReceiptsBar({Key key}) : super(key: key);
-
-  static ReceiptSortTypeEnum selectedReceiptsSortType =ReceiptSortTypeEnum.cost;
-  static bool isIncreasing=false;
-  static String selectedSortName;
-
-  @override
-  _SortReceiptsBarState createState() => _SortReceiptsBarState();
-}
-
-class _SortReceiptsBarState extends State<SortReceiptsBar>
-    with TickerProviderStateMixin {
-
-  Animation _arrowAnimation;
-  AnimationController _arrowAnimationController;
-
-  Map<String, ReceiptSortTypeEnum> nameValueMap = {
-    "Cost": ReceiptSortTypeEnum.cost,
-    "Company": ReceiptSortTypeEnum.company_name,
-    "Category": ReceiptSortTypeEnum.category_name
-  };
-
-  @override
-  void initState() {
-    super.initState();
-    _arrowAnimationController =
-        AnimationController(vsync: this, duration: Duration(milliseconds: 300));
-
-    if(!SortReceiptsBar.isIncreasing){
-      _arrowAnimation =
-          Tween(begin: 0.0, end: pi).animate(_arrowAnimationController);
-    }
-    else{
-      _arrowAnimation =
-          Tween(begin: pi, end: 0.0).animate(_arrowAnimationController);
-    }
-
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(children: [
-      Container(
-        color: Colors.white,
-        child: DropdownButton<String>(
-          hint: Text("Select item"),
-          value: SortReceiptsBar.selectedSortName,
-          onChanged: (String value) {
-            setState(() {
-              SortReceiptsBar.selectedSortName = value;
-              SortReceiptsBar.selectedReceiptsSortType = nameValueMap[value];
-              Navigator.of(context).popAndPushNamed(HomeScreen.tag);
-//            HomeScreen.refreshFunc();
-            });
-          },
-          items: nameValueMap.keys.map((String sortName) {
-            return DropdownMenuItem<String>(
-              value: sortName,
-              child: Row(
-                children: <Widget>[
-                  SizedBox(
-                    width: 10,
-                  ),
-                  Text(
-                    sortName,
-                    style: TextStyle(color: Colors.black),
-                  ),
-                ],
-              ),
-            );
-          }).toList(),
-        ),
-      ),
-      AnimatedBuilder(
-        animation: _arrowAnimationController,
-        builder: (context, child) => Transform(
-          alignment: Alignment.center,
-          transform: Matrix4.rotationX(_arrowAnimation.value),
-          child: IconButton(
-            icon: Icon(Icons.sort),
-            onPressed: () {
-              setState(() {
-                if (SortReceiptsBar.isIncreasing) {
-                  SortReceiptsBar.isIncreasing = false;
-                } else {
-                  SortReceiptsBar.isIncreasing = true;
-                }
-
-                _arrowAnimationController.isCompleted
-                    ? _arrowAnimationController.reverse()
-                    : _arrowAnimationController.forward();
-
-                Navigator.of(context).popAndPushNamed(HomeScreen.tag);
-//                HomeScreen.refreshFunc();
-              });
-            },
-          ),
-        ),
-      ),
-    ]);
-  }
 }
